Quality of life changes for winjul21aud
These changes are for winjul21aud. They increase the quality of life for winobj by removing the need to
run winobj in compatibility mode (aka - giveio issues). They will also only prompt the user to install
giveio if it is necessary for the experiment. Regular experiments using the keyboard and mouse do not need giveio.